requirements.txt
Flask==3.1.2
python-dotenv==1.1.1
google-cloud-storage==3.4.0
